Moved tests to new file and test for correct orderID after amends
